+++ tests/game_hhc/src/h/input.h
@@ -44,6 +44,7 @@ extern fsl_key_bind bind_zoom;
 extern fsl_key_bind bind_toggle_flashlight;
 /* extern fsl_key_bind bind_pause; */
 /* extern fsl_key_bind bind_chat_or_command; */
+extern fsl_key_bind bind_reload_shaders;
 
 /* ---- debug --------------------------------------------------------------- */
 
@@ -61,7 +62,6 @@ extern fsl_key_bind bind_toggle_chunk_bounds;
 extern fsl_key_bind bind_toggle_bounding_boxes;
 extern fsl_key_bind bind_toggle_chunk_gizmo;
 extern fsl_key_bind bind_toggle_chunk_queue_visualizer;
-extern fsl_key_bind bind_reload_shaders;
 
 /*! @brief setup all key binding values and rules.
  */